requirements.txt
appdirs==1.4.4
astroid==2.4.2
click==7.1.2
dataclasses==0.7
importlib-metadata==1.7.0
isort==5.4.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
mypy-extensions==0.4.3
mysql-connector-python==8.0.21
pathspec==0.8.0
protobuf==3.13.0
pycodestyle==2.6.0
pylint==2.6.0
python-dotenv==0.14.0
regex==2020.7.14
six==1.15.0
tabulate==0.8.7
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
wrapt==1.12.1
yapf==0.30.0
zipp==3.1.0
